Add cuda implementations for unary and binary tensor operations in #341 and #334 (#346)
* Add cuda implementations for unary and binary tensor operations

* Add cuda kernel for powi; Use fewer 64-bit functions

* use copysign in abs kernal, as suggested in https://github.com/coreylowman/dfdx/issues/341#issuecomment-1376392725

+++ src/tensor_ops/exp/exp.cu
@@ -10,7 +10,7 @@ extern "C" __global__ void exp_forward(
     if (i >= numel) {
         return;
     }
-    out[i] = exp(inp[i]);
+    out[i] = expf(inp[i]);
 }
 
 extern "C" __global__ void exp_backward(
@@ -24,5 +24,5 @@ extern "C" __global__ void exp_backward(
     if (i >= numel) {
         return;
     }
-    grad_inp[i] += exp(inp[i]) * grad_out[i];
+    grad_inp[i] += expf(inp[i]) * grad_out[i];
 }
